Ml-model aided by rule_based and lookup
I've created a simple ML-model which I have trained on my data. I have further improved the categorization with a lookup-table and a rule-based system. If no prediction is found a prompt for the use to add manually appears. Only in console at this moment but it was made just to see how it works.